Fix some swap chain issues
- Add image acquisition in `VulkanSwapChain::WaitForFrame()`.
- Fix frame index counting in `D3D12SwapChain`.
- Upgraded D3D12 swap chain from version 1 to 3.
- Removed unused member function `ISwapChain::GetApiHandle()`.
- Switch from `uint64`->`uint32` for frame index counter.
- Rename some things.

+++ Render/Core/ISwapChain.cpp
@@ -13,13 +13,13 @@ ImageBuffer* ISwapChain::GetCurrentBuffer()
 	return &buffers[currentFrameIndex % numBuffers];
 }
 
-uint64 ISwapChain::GetSignalValue()
+uint32 ISwapChain::GetSignalValue()
 {
 	uint32 bufferIndex = currentFrameIndex % numBuffers;
 
 	frameCounters[bufferIndex] = currentFrameIndex;
 
-	return (uint32)frameCounters[bufferIndex];
+	return frameCounters[bufferIndex];
 }
 
 ISwapChain::ISwapChain()
